docs: (1/n) move docs in their own folder
