+++ src/docs/components/highlight-chars.tsx
