+++ src/docs/views/basic.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 
-import { Fzf, FzfResultEntry } from "../lib/main";
+import { Fzf, FzfResultEntry } from "../../lib/main";
 import { HighlightChars } from "../components/highlight-chars";
 import wordList from "../lists/words.json";
 import dateFnDirList from "../lists/date-fns-repo-folders.json";
